ftraceparser: Add skeleton for prescan
This preScan() function in the FtraceParser class will do the
initial scan of the events to determine at least the following:
maximum number of CPUs, number of events, start time and end
time. The determination of the number of CPUs need to be
improved so to see if there are higher number CPUs appearing
as params of certain events, such as cpu_frequency or
sched_migrate_task.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>