ftraceparser: avoid doing double QMap lookups
This should avoid the unfortunate sequence that we first lookup
whether there is a task with a certain pid, and then depending in the
answer either lookup the value or insert a new task.

I believe that this function will be one of the hottest functions
in the whole application. In the future, it may make sense to
optimize this further, possibly by getting rid of the QMap
alltogether and try to select a faster data structure.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>