ftraceparser: Add a riser to the incoming edge of a task
When a task is scheduled there will be a riser, the height will be
full if the task has waited at least 20 ms from wakeup time to the
time when it's scheduled. If the task was previously scheduled out in
running state, then that will be regarded as the wake up time.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>