Replace ParserThread with more generic WorkThread
This new WorkThread class, which is a template class, can be used to
call member functions in any class. I am not 100% convinced about the
usefulness of this approach. One downside is that all threads being
run through this framework will get the name __TThread, which is kind
of boring.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>