Add CursorInfo
This adds the CursorInfo class. It's widget with an editable line of
a double value and a push button. It serves a dual purpose, informing
the user of the current time value of a cursor and with the button
gives the user the possibility of moving the cursor to a specific
time value.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>