MainWindow: Improve cursor setup and scroll eventsview to the red position
This modification will improve the cursor setup and scroll the
eventsview ot the posotion of the red cursor. There is some
weird behavior of the EventsWidget, it will not agree to scroll
to be immediate scrolled the middle of the trace when the trace
has just been setup, it needs to be scrolled around a bit first.
This weird behavior should be investigated at some point but for
now, let's have a workaround.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>