MainWindow, FtraceParser: Use adaptiv sampling for scheduling graph
This is the real way of implementing adaptive sampling. QCPCurve has
to be replaced by QCPGraph. Unfortunately, this also means that
the we will lose the rendering of wakeup time and the case when
task is still running while being scheduled out. However, I think that
it's better to use QCPGraph here and try to show those other things
with complimentary QCPGraphs, since the benefit from using QCPGraph
is huge, thanks to it's adaptive sampling.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>