MainWindow, FtraceParser, Task: Finalize adaptive sampling change
This will graphs that display wakeup time and still running
information, which was removed when adaptive sampling was taken into
use. Running time is for now rendered by a blue triangle at the rising
edge of the scheduling, the height of the triangle signals the wakeup
time. If the task at the bottom of the rising edge has a red circle,
it means that the task was still running when it was scheduled off
the CPU.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>