FtraceParser: Improve estimation of wakeup time
This patch assumes that a CPU will not enter idle, if it has a task
that has been woken up, that is if it seems to have happend, we
assume that the there is a missing wakeup event somewhere.

Another guess used is that for new tasks, it's assumed that if
the CPU has been idle, then it's not the case of new task
in the beginning of trace, where the wake up occured before
the trace begun, instead we assume that there is a missing
wakeup event.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>