ftraceparser: Implement correction of kernel side rollover bug
It happened, that a trace had the following series of timestamps:

31548.999999
31548.999999
31548.999999
31548.1000000
31548.1000000
31548.1000000
31548.1000000
31548.1000000
31548.1000000
31548.1000000
31548.1000000
31548.1000000
31549.000000
31549.000000
31549.000000

Clearly, this is a bug in the kernel where the decimal parts overflows
while the whole number of seconds have not yet been updated. The
correction for this is to check if a timestamp is smaller than the
previous timestamps, and if it's the case, then add 0.9 to the
timestamp.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>