TraceFile: Allocate from one pool per ThreadBuffer
Here we reaplace the single pool with a number of pools. The idea
behind this is that we can have one pool per ThreadBuffer (in
FtraceParser). This way, we will be able to clear the buffer if we
have a consuming thread that will use the StringPool mechanism to
allocate memory for Events, because then the necessary strings
will be allocated by the string pools and copied there as necessary.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>