Add the StringPool mechanism
A stringpool is a special string allocator that is used to allocate
strings in the case when there are many strings that are identical.

It works in the following way: the StringPool::allocString() is called
it passes a (temporary) copy of the string of concern. If it can find
that an identical string has already been allocted, it merely returns
a pointer to the previously allocted string. If not, it allocates
a new string and copies the content of the temporary string to the
newly allocated string and returns a pointer to the newly allocated
string.

The internal datastructures consist of hash table of binary search
trees, unfortunately unbalanced. The fact that the trees are
unbalanced makes it practically impossible to read large traces with
contains events with ascending or descending numbers, since the
monotonicity of the numbers will hit the worst case of the unbalanced
tree.

This would be possible to fix by improving the hashing function and
possibly expanding the hash table, and/or by making the binary trees
balanced, such as RB trees. Also, it would be possible to use some
heuristic rules, for example one heuristic could be to count the
number of inserts at a particular hash value and if it's large enough
when inserting, counting the number of digits in the string, for
example two or more digits could disqualify the string. A disqualified
stringwould then be allocated straight from a mempool with no attempt
at reuse.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>