Source file of the first variant of the shark icon
This was the first incarnation of the shark icon. A png of this was
never commited but I will commit it first, and then all subsequent
version of the icon until the one that matches the current shark.png.
I do this in order to store the complete development history of
Traceshark for posterity.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>