Add StringTree to determine event type and reduce strcmp() use
This will introduce a new class StringTree that is very similar to the
already existing class StringPool. The main difference is that it is
able to store a number during the first lookup of a string, this
number will be retrieved when doing subsequent lookups.

This enables us to do an essentially cost free determination of event
type when alloacating strings, since StringTree also has the string
allocation properties that StringPool has.

StringTree is used by the EventNode class.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>