StringTree: Separate the tree lookup into a while() loop
This does the same for StringTree as was previously done for
StringPool, so I copy paste the commit message here:

This will put the tree lookup part into a separate while() loop,
instead of having the insertion code there in the middle, with a
goto loop.

I think this separate while() loop can be better optimized by the
compiler, also it gives us the option to try manual unrolling of
the loop. Initial performance testing seems to confirm that the
compiler is able to generate faster code this way.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>