StrHash32(): Simplify hash function
This makes the hash function simpler and faster, in addition it has
the property that strings that start with the same 4 chars will get
the same hash value; this is maybe not so great but it might have a
nice interaction with the recent string comparison optmization in
StringPool where we are avoiding to do repeated comparisons of the
first N characters if those have already been compared. This will
result in strings that start with the same four characters ending
up in the same three and then the StringPool optimization can kick
in for at least those 4 characters.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>