TaskGraph: Store a pointer to itself in the legend graph
The legend graph gets its own clas LegenGraph that contains a pointer
to the TaskGraph that created it and to which it's associated.

This will be required when we add interactions regarding clicking and
double clicking on legend items, because then we will only have a
pointer to the LegendGraph object.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>