MainWindow: remove legend items when double clicked
This patch enables us to remove items from the legend by double
clicking on them. There is a crazy caveat though, since double
clickin is also the interaction for moving the cursor, it will
also move the cursor to the place in the legend where you are
double clicking. This is a bit annoying and should be fixed
later, either by filtering the moving of cursors so that we don't
move the cursor if the double clicking hit a legend item; this
may be difficult to achieve so some band aid solution may be
necessary, such as making it possible for the user to have a third
state in the active cursor selector: none.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>