MainWindow: add filtering to double clicking on plot
This modification adds filtering to the function for handling double
clicking so that the cursor isn't moved when the double clicking hits
a legend or legend item. The crazy thing here is that I have to
introduce a new derived class of QCustomPlot only because
QCustomPlot::layerableAt() is a protected function. I don't see any
other alternative because I need to use the function and I don't want
to modify QCustomPlot as long as it isn't a clear bug.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>