New feature: Add the display of stack traces
This chang makes it possible to display stack traces by double
clicking on the on a given event's "Info" column in the events
view. This double clicking will open a dialog that displays the
stacktrace for that event.

Of course, this requires that the perf trace has been recorded with
the -g flag to capture stack traces.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>