Improve naming
This will change EventDialog to EventInfoDialog, since it is a dialog
that will open when you click on the event info field. In the future
we might have a dialog that opens when you click on the event field
and then it will be nice if the name is not reserved.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>