TList: introduce TList
TList is an alternative to QList. It doesn't have all the features
of QList but on the other hand is has preallocation, which should
gives us a performance boost in the parsing functions.

It's written as a template but it was written with the
QList<TraceEvent> events in mind that exists in the TraceParser.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>