TList: try to make the code a bit more portable and robust
The idea here is that the TList should work even if UINT_MAX is more
than 0xffffffff, for example if it's a 64-bit value. Also, there is
some code to make TList work on a platform with 16-bit unsigned ints
but that is probably moot.

I am not really sure about this nullptr business, is it really appropriate
to give as an argument to mmap(), I hope so.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>