Fix bug with regards to task name containing a space
Altough the parsing code is quite well protected against spaces in the
arguments of an event, it isn't at all protected against a space
in the task name field, and this concerns both ftrace and perf events.

This change, modifies the parsing quite heavily, it introduces the
following hack:

The GrammarNodes that earlier was parsing the name and pid, will now
just store those temporarily in the the argv[] field of the
TraceEvent. The node for CPU number will both parse and store the
string to argv, then when the timenode is parsed, it's "time
for choosing", then it's checked if the temporarily stored
arguments matches name, pid and time, if it does, then the argv
vector is reset and everything stored. If it doesn't match, then
the grammar tree has been set up so that it will assume that
we are still in some split up part of the task name and there
will be a retry.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>