New Feature: Add button for finding wakeup events
The idea here is to implement a button that finds the previous wakeup
event for the currently selected task. The starting point is the
currently active cursor:
- First the preceding scheduling switch event is found, relative to
the cursor, for the selected task
- Second, the wakeup event that corresponds to the scheduling switch
event is found.
- The active cursor is moved to the sched switch event and the inactive
cursor is moved to the wakeup event.
- The events view is scrolled to the wakeup event.

At least this is the idea; beware of bugs :)

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>