MainWindow, TraceParser: Improve the "Find wakeup" functionality
This will make the "Find wakeup" functionality yet a bit more
robust by using indexes for the found scheduling event instead
of timestamps, I guess this would mainly be useful if there
would be a situation with wakeup and sched_switch event with
the same timestamp or something like that. I think it wouldn't
happen on the x86 architecture that I myself am using but
I am not sure about all other architectures out there.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>