TraceParser: Huge refactoring of perf and ftrace code
The goals behind this reafactoring is the following:
- Eliminate unecessarily duplicated code between ftrace and perf
- Make the code more readable
- Don't sacrifice any performance.

To avoid the performance loss, the code is duplicated in many places
by using __always_inline and a function argument of type tracetype_t
that tells whether it is perf or ftrace, so that the relevant
switch() and if() statements can be evaluated at compile time. Or to
put it more simply, we let the compiler duplicate the code instead
of duplicating it in the source files.

This is a somewhat large and risky commit, beware of bugs, but it's
necessary for the long term viability of traceshark.

It can also be mentioned that some bugs where corrected during the
refactoring, for example, the problem that short traces would not
get displayed at all.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>