TraceParser: cosmetic include change
One of the ideas behind the recent refactoring of TraceParser was that
it should not directly use the functions in ftraceparams.h or
perfparams.h but instead use the generalized functions defined in
genericparams.h.

This change is purely cosmetic since inlining is used, genericparams.h
will include both of the other files and traceparser.h already
includes genericparams.h.

However, I believe that impressions matter :)

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>