TraceAnalyzer: Break out parser code and put it in separate class
This is the second part of the great refactoring of the TraceAnalyzer
class, which was earlier was called TraceParser.

This change moves the parsing code from TraceAnalyzer and put it in a
new class that is called TraceParser, which is a member of
TraceAnalyzer. By separating the parsing code from the analyzer code
it's easier to understand and maintain the code.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>