ThreadBuffer: Use QWaitCondition instead of double QMutex
The problem with the previous implementation of this is that locking
and unlocking a mutex from different threads are not officially
supported. It happens to work but it relies on undefined behavior.

This is an attempt to reimplement the synchronization functions in
the ThreadBuffer class so that it relies on behavior that is more
well defined.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>