TraceParser, TraceAnalyzer: Improve performance
This change skips the prescanning functions. Instead, it is assumed
that nobody will have more than 256 CPUs. This has the benefit that
the parsing and processing (i.e. analysis) part can be run in
parallel.

Thus, now loading, line reading, parsing and analysis of the trace
is all done in parallel in a pipelined fashion.

This change has a significant performance improvement if you have
more than two CPU cores, around 13% for my sixcore machine.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>