Use prctl() to give names to threads
This change will cause prctl() to be run to set a name for the
threads. The threads of the WorkQueue will now be called
WorkThread, there will be readerThread, parserThread and
LoadThread. The analysis will be run by the main thread
so that will not have a separate name.

The motivation for this change is just to make it easier to
use traceshark to analyze the behavior of itself.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>