Improve the name detection of threads that changes name
Earlier the name displayed when selecting a task was derived as
folllows:
- it was the first name detected by the parameters of a switch event
on that particular CPU. This meant that when clicking on the task on
different CPUs, different names would be displayed.

If considering the "typical" case for how tasks changes name, it goes
somewhat like this:

1. fork or clone a new task
2. Run for a while
3. Change name by using prctl() or similar method.
4. Exit the task.

Now, after this change the names that a task has will be put into
a linked list in the Task class that will be used as a stack. The
most relevant name will be the one that was last seen during the
whole trace and the same name will be displayed regardless of
which CPU you click on. In the future, it would be possible to
implement a functionality that display previous names inside
parentheses or something.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>