TraceAnalyzer: Make task name detection more accurate
This should increate task name detection accurazy, so that any name
that is used in conjunction with sched_switch events will be recorded,
with one exception: if there are new names that only appear in the
parameters of the sched_switch event. I am not sure if this could
ever happen but I believe that it could only happen in some idiotic
case towards the very end of a trace.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>