Add the displaying of task name history in the legend
If you add a task to the legend with the "Add to legend" button, the
name used to be displayed as "last name:pid", now with this change
it will be displayed as "last name(previous name):pid" if there was
a previous name, if there were more than one previous name three
dots "..." will be appended to the end. How ever, the task info that
is displayed when you click on a task will still only have the last
name, because I feel that there isn't enough room in the UI to
have previous names there.

This change is merely useful, when you are not as sure of which tasks
are related. Sine the previous name is isually (I believe always?)
the same as the name of the task that forked it. Of course, this only
helps if the tasks is forked during the trace.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>