MainWindow: Fix regression regarding cursors drawn under graphs
When the setupCursors() was reshuffled in the previous commit it had
the unintended consequence that the cursors where drawn underneath
the graphs, which to me looks kind of crazy.

This commit fixes the problem with the help pf QCPLayer class,
creating a separate cursor layer that is defined as being on
top of the main layer where the graphs are drawn.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>