MainWindow: Fix second regression from moving setupCursors()
This fixes the second and hopefully last regresson that came when I
moved the setupCursors() function to be executed earlier for
performance reasons.

The problem was that the events widget did not scroll to the active
cursor position when the trace was opened. The reason was simply
that the EventsWidget class had not been configured with the
events list, so it could not scroll when setupCursors() was run.

Now this change makes sure to set up the evetns before
setupCursors is being run.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>