TaskInfo: Replace the text on "Find wakeup" button with icon
This introduces the concept of buttons with icons instead of text.
This is desirable in order to save space in the UI. I don't care
about the looks and I con't kid myself to think that the icons
that I am able to draw would be more aesthetic than the text.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>