MainWindow: Fix third regression from moving setupCursors()
Unfortunately, I discovered a third regression from commit
46cb72ab3c6c5ed98937d613686d94b4175a104e. This is caused by
the fact that the setupCursor() code calls Cursor::setPosition()
which calls TracePlot::replot(). Earlier this was done after
MainWindow::showTrace(), so it was guaranteed that there
would be a replot after the graphs had been added to the plot. This
is no longer true, so the solution is to add a replot() call
at the end of MainWindow::showTrace()

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>