New feature: Add unified taskgraphs
This is commit introduces two new buttons which makes it possible for
the user to add and remove unified task graphs. Unified means that it
is a union of how the task runs on all the CPUs.

It's rendered identical to the per CPU taskgraphs, except that the
wakeups are displayed horizontally, instead of vertically. I wanted
to do this horizontal thing already for the per CPU task but there
it's not such a good idea because in a situation where tasks have to
wait for a long time, it could happen that wakeups would overlap and
it would look like a mess.

However, here with these unified taskgraphs, we do not have the
problem of possible overlap because all unified TaskGraphs occupies
their own exclusive vertical space.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>