EventsWidget: Fix bug about application not working if events view is hidden
It's possible for the user to show/hide the EventsWidget by right
clicking in any of the docking areas and (un)select Events.

However, up until now there has been severe misbehavior if doing that
and either trying to load a trace or move a cursor. This patch should
fix the issue. It's apparently caused by the fact that it's not
advisable to call TableView::resizeColumnsToContents() if the widget
isn't visible.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>