MainWindow: Bugfix for showWakeup()
When the "Find wakeup" button is pressed, the active cursor is moved
to the time when the task is scheduled and the inactive cursor is
moved to the preceding wakeup event, and the events view is scrolled to
the wakeup event. The inactive cursors time is updated in the info
field to match where it has been moved but the active cursor's time
isn't updated.

Thie commit fixes the bug so that the active cursor's time is updated
to the value that corresponds to the place where the cursor has been
moved.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>