Bugfix: Prevent Task::checkName from being called with nullptr
The implementation of the Task class assumes that it's checkName()
function is never called with nullptr as an argument. I believe that
this is a sound assumption but unfortunately currently the
TraceAnalyzer doesn't check for nullptr before calling checkName.
This is a bug that can happen in practice and it causes the whole
application to crash with SIGSEGV.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>