Improve zoom range adjustment when adding/removing a task graph
When a unified task graph is added/removed the upper zoom range was
needlessly adjusted before this change. Now, we only adjust the lower
part so that the task becomes visible or, when the task is removed
it's adjusted so that any freed up superfluous white space is zoomed
out of view.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>