TraceAnalyzer: Add heuristic for detecting when wakeup events are missing
It seems that in some cases processes will not have wakeup events,
particularly the perf process.

With this change, we will keep track of when a process went to sleep, if
the the last wakeup time is older than the last time when the process
went to sleep, then we will not draw any wakeup time at all.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>