TraceParser: Put some members into a new composite class
The purpose of this change is to smooth the way for the implementation
of a new feature. I hope to be able to support Ftrace backtraces also.

In addition, I feel that this is more clean to have separate instances
of these variables for Ftrace and perf, since in the beginning of the
trace, both parseLinePerf() and parseLineFtrace() will be called.
Currently, it doesn't seem to be any interference between them but it's
not very easy to see from the code and in the future, it would be very
easy to make such changes without considering the fact that both
functions relies on the very same variables.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>