TraceAnalyzer(): Prevent absurd wakeup times from being drawn
handleWrongTaskOnCPU() tried to adjust the lastWakeUP time in an effort
to make it closer to the truth. While this may be true it also prevents
the estimateWakeUp() from detecting the absurdity of the situation.

Now instead we adjust the lastSleepEntry, so as to guarantee that the
absurd wakeuptime will be discarded in case of lost events.

This still doesn't cover all of the cases of incorrect wakeup times in
case of lost events but it prevents us from doing the wrong thing when
we have enough information to do a better estimate.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>