Avoid unnecessary copying of all data between buffers
This is an optimization of the parser so that it will not copy all the
data in the LoadBuffers to MemPools, instead the ThreadBuffers will
contain pointers to the data in the LoadBuffers, so that the
LoadBuffers must not be reused before the ThreadBuffers have been
consumed. This is achieved by the ThreadBuffer maintaining a pointer
to the corresponding LoadBfuffer and giving the LoadBuffer class three
significant states that are synchronized between threads:
LOADSTATE_EMPTY, LOADSTATE_LOADED, and LOADSTATE_TOKENIZED. Before
this change, there were only two states: isEmpty could be true or
false.

This complicates the code a bit but there is a large performance gain.
After this change, it appears to be that the bottleneck has shifted
away from the readerThread to the parserThread. I believe that it may
be possible to achieve even more performance gains if the parserThread
could be optimized, which it probably can.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>