This is a rewrite of the Ftrace and Perf grammars
It seems that this rewrite improves the performance a bit by
abandoning the object model with an explicit grammar tree and
instead using a switch-case structure.

There is also an improvement of the TList class. I have added a
softclear() function that will not give back any memory to the kernel
but instead just adjust the number of allocated elements to zero while
keeping the memory around. This is used in situation where we are
periodically filling up and clearing a TList during parsing, such as
in the ThreadBuffer class.

Finally, in order to capitalize on the other changes, the numbers of
load buffers and thread buffers are increased to 4.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>