TraceParser: Fix bug with garbled task names
This bug was introduced when the ftrace and perf parsers were
rewritten because the new parsing code doesn't reset the
event.argc variable after parseLineFtrace() has failed in
TraceParser::parseBuffer(), which it always will wen we are parsing
a perf trace. This means that the name of the event will be
incorrectly determined to be the whole line of the event.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>