Ftrace: Fix issue with double names in the Legend
This fixes an issue that mosly effects the Legend. Where seemingly
identical task names shows up twice, as both new and oldname. The
cause is that the Ftrace parser code incorrectly adds a leading
space to the task names of events. This causes the task names to
be different from the task names that are extracted from the
parameters of events, such as wakeup and switch events.

The result of this is that the same name might be registered as
a task that has changed names, because of the leading space.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>