Sanitize the Ftrace parser code a bit
This will get rid of a superfluous if clause, the
"if (namestr.len != 0)" test can basically be merged with the
preceding "if (event.argc > 2)", since event.argc > 2 implies
that we will put things into namestr.

I am not aware that this would fix or introduce any bug. It
could possibly be a bit faster but I doubt it. The main thing
is that it makes the code easier to read, at least for me.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>