TaskModel: Improve the taskmodel
This should improve the scrolling performance of the TaskSelectDialog,
when there are many tasks with the same name. With this change, an
ordered TList of pointers to Task objects will be created when the
setTaskMap() function is called. Thus when data is requested for
individual cells, it can be resolved by looking up this list.

Also, since the sorting of the tasks is dong "offline", the bubble
sorting is done to completion, so now the tasks with the same name
are ordered by pids.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>