TraceEvent: Get rid of the eventName member
This is performance optimization which tries to exploits the
following things:

* The evenName field is only needed by the EventsModel class for
  displaying event names in the EventsWidget.

* The analyzer code already only uses the type field to distinguish
  between different types of events.

* It's cheap to create a system for looking up event names from
  the type field.

* Removing a pointer from TreaceEvent will probably give a small
  performance boost and also save some memory.

In order to be able to still display the names, we arrange for the
StringTree class, used by the parser to construct a lookup table
that maps event types to pointers to TString, these are the TStrings
that are allocated by the searchAllocString() during parsing.

We save a pointer to the StringTree that was used for parsing in
a static member of the TraceEvent class. Thus the new getEventName()
function in TraceEvent can lookup the event name from the type.

In addition, we need to have in the parsers a counter for unknown
events, in order to be able to give unique type values for every
encountered unknown event.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>