PerfGrammar: Add support for parsing some more events
There is a type of perf event that has not so far been possible to
parse correctly, for example:

swapper     0 [001] 24423.536338:          1 instructions:

There are many more vents of this type, they are among other things
various HW events, such as cycles, cache misses and so on.

With the above example, the problems with the old code are that:

* The parser gets confused with the number before the event

* The parse gets conused by the fact that the event name only has
  one colon at the end and no colon in the middle.

With this patch, it's possible to parse these events and the value
of this numerical extra parameter is added to a new field in the
TraceEvent class. When viewing events with the events widget, this
number is displayed as if it had been the first argument parameter
of the event.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>