CursorInfo: toolbarize the move cursor button
This change will change the QPushButton, the one that could be used to
move the cursor to the specified value, into a button in a toolbar.
This will serve to unify the size of the buttons, so that all buttons
become toolbar buttons and of the same size and spacing.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>