Parser: Increase the fallthrough in grammar switch statements
This change is a performance optimization that is not meant to change
the behavior in the code in any way. The idea is to use fallthroughs
in the switch statements as much as possible, instead of setting the
state variable to the next state and doing another loop iteration.

This has a measureable speedup, at least on x86_64.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>