Implement a complete parsing of task state flags
Previously, the parsing of the task state flags was done so that the
parsing  in perf_sched_switch_state() and ftrace_sched_switch_state()
functions mereley tested the first character to see if it was equal
to 'R' and if it was, then the task is deemed runnable.

This new implementation is able to parse all of the flags that can be
emitted by the sched_switch event, at the time of writing, referring
to version 4.7.2 of the Linux kernel.

Currently, the information contained in these extra flags is not used
by traceshark but I am considering to in the future implement
visualiztion of preemption and uninterrutple sleep.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>