New feature: show preempted graph
Previously, there was a red circle drawn when a task in the running
state scheduled out, regardless of whether the task was preempted or
not.

Now with this change, a task in the running state will get:
- a red disc if it was preempted
- a blue disc if it was not preempted

Tasks that go to sleep will not get any extra decoration, regardless
of what kind of state flags they have, just as it was earlier.

Another way to look at this is that that the red circles has been
replaced with blue nad red discs. The color of the disc tells if
the task was preempted.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>