MainWindow: Change the discs back to circles
Those discs that are marking that a task is still running when
scheduled out, is now changed back to circles again. I believe that
circles makes it easier to see whether there are a lot of them close
together or not. This is nice when you are zoomed out and want to get
an overview.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>