Add const where it makes sense
I have been a bit neglient with const. There have been a lot of places
where it could have been added. Especially after member functions to
declare that they do not modify the (this) object.

This modification adds const to at least some of these places. I am
not able to say if there is any performance benefit from this, if it
there, then it's a little bit too small to easily measure with
adquate confidence.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>