AbstractTask: Make sure that all scalar members are initialized
This change adds pid, offset and scale to the initializer list,
although I think that this will have no impact, these members
should get their real values during the analysis of the trace,
before they are consumed but let's fix it anyway.

This issue was found by Coverity. It's CID 1358234 in Coverity.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>