ThreadBuffer: Initialize the loadBuffer pointer to nullptr
The real value is not known when the constructor is run so use nullptr
for initialization. The impact of this defect is zero but let's fix it
anyway so that we can be sure that all members are initialized. It's
anyway better that, should a user of this class by mistake forget to
set the loadBuffer pointer that it's nullptr than some arbitrary
value.

This issue was found by Coverity. It's CID 1358225 in Coverity.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>