MainWindow: New feature - make it easier to iterate through wakeups
This modifies the behavior of the "Find wakeup" button. The old behavior
was to move only the cursors, so that the active cursor was moved to the
"scheduling edge" and the inactive to the wakeup event.

With this, the cursors move the other way around, the active to the
wakeup event and the inactive to the scheduling edge.

Assuming that the wakeup event did not happen during pid 0 (swapper), then
the task that executed during the wakeup is selected and the previou task
is deselected. The idea is that the task that was executed during the
wakeup is a "wakeup culprit candidate", not necessarily true because the
wakeup might have been caused by an interrupt.

If pid 0 was executing, then it's a clear case of an interrupt and the
selections do not change, the old task is still selected.

The idea is that the user can determine whether the interrupt was
caused by the candidate by looking at the backtrace.

A user who wants to trace how some activity propagates through
several task would do something like the following iteration:

1. Select the task of interest and move the active cursor to the
   point of time that is of interest.

2. Click the "Find wake up" button, i.e. the alarm bell button.

3. Verify that a wakeup event has been found (event view should be
   scrolled to an wake up event).

4. Double click on the wake up event's Info field in the event view.

5. Inspects the backtrace and check whether it is and irq, syscall, or some
   special case. IRQs and syscalls should be the most common.

6. If it's an IRQ, then the investigation ends. A new task might be selected
   but it is not the wake up culprit.

7. If it's a syscall, then a new task should have been selected (if not,
   something went wrong, such as missing events).

8. He can then press the "Find wake up" button again because the active
   cursor is now at the wakeup event and the canidate task has been
   selected. This should find the wake up event that woke up the task that
   woke up the original task.

9. Go back to step 3.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>